data/requirements.txt
mysql-connector-python==2.0.4
numpy==1.11.3
pkg-resources==0.0.0
